requirements.txt
flask==3.0.0
flask_sqlalchemy==3.1.1
flask_migrate==4.0.7
psycopg2-binary==2.9.9
flask-jwt-extended==4.6.0
flasgger==0.9.7.1
python-dotenv==1.0.1
pytest==8.2.2
Werkzeug==3.0.3
